+++ deuda-service/src/main/java/com/jonathan/rest/deudaservice/repository/DeudaRepository.java
@@ -5,6 +5,6 @@ import com.jonathan.rest.deudaservice.entity.Deuda;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface DeudaRepository extends JpaRepository<Deuda,Long>{
+public interface DeudaRepository extends JpaRepository<Deuda,Integer>{
     
 }
