+++ deuda-service/src/main/java/com/jonathan/rest/deudaservice/repository/EstadoRepository.java
@@ -4,6 +4,6 @@ import com.jonathan.rest.deudaservice.entity.Estado;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadoRepository extends JpaRepository<Estado,Long>{
+public interface EstadoRepository extends JpaRepository<Estado,Integer>{
     
 }
